requirements/base.txt
Django==2.2.24
wagtail==2.13.2
celery==4.2.1
django-compressor==2.2
django-environ==0.4.5
django-foundation-formtags==0.1.0
django-wagtail-feeds==0.1.0
django-libsass==0.7
django-redis==4.10.0
libsass==0.16.1
psycopg2-binary==2.8.5
wand==0.5.0
bleach==3.1.5
ls.joyous==0.9.0
wagalytics==0.8.1
django-countries==5.3.3
django-allauth==0.39.1
django-braces==1.13.0
django-mptt==0.11.0
django-leaflet==0.26.0
django-publications==1.0.0
idigbio==0.8.5
xlrd==1.2.0
django-import-export==2.0.2
django-ckeditor==5.9.0
django-map-widgets==0.2.3
unicodecsv==0.14.1
wagtailgeowidget==5.0.2
lxml==4.6.3
pygeoif==0.7
dateutils==0.6.7
Pyzotero==1.4.21
pyshp==2.1.3
mdb-parser==0.0.2
django-geojson==3.2.0
